Reduce Unload Queue Complexity
Revert TacoSpigots change to use a ArraySet.

An ArraySet is not the correct collection for this job. It provides o(n)
add and remove it also provides fast iteration which is severely limited
to 100 actions in this case. Chunk Loading and Unloading modify this set
which can be a hotspot when theres a lot of players loading chunks.