Column TNT Spawning
This significantly improves TNT Merging and Movement Caching efficiency,
in-game through testing this offered an additional improvement of ~1.5x
to the entity count while firing a cannon with TNT Merging enabled.

Though this patch can affect cannons, it can mostly be overlooked as
most issues caused would also be affected by the redstone algorithm used
by the server causing the cannon to be unreliable.

As this does cause some issues, even if they are minor it is disabled by
default.